+++ Trainer_v4/Constants.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Trainer_v4
+{
+    public static class Constants
+    {
+        public const int ROWS_TO_SKIP = 7;
+
+        public const int ELEMENT_WIDTH = 150;
+        public const int ELEMENT_HEIGHT = 32;
+
+        public const int Y_WINDOW_OFFSET = 260;
+        public const int X_WINDOW = 670;
+
+        public const int FIRST_COLUMN = 1;
+        public const int SECOND_COLUMN = 161;
+        public const int THIRD_COLUMN = 322;
+        public const int FOURTH_COLUMN = 483;
+
+        public const int FIRST_ROW = 0;
+        public const int SECOND_ROW = 32;
+        public const int THIRD_ROW = 64;
+        public const int FOURTH_ROW = 96;
+        public const int FIFTH_ROW = 128;
+        public const int SIXTH_ROW = 160;
+        public const int SEVENTH_ROW = 192;
+    }
+}